refactor(web): unify ACL and Forward rule-counter hooks
